+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/support/stretching/AbstractStretchingScreen.java
@@ -5,6 +5,7 @@ import com.refinedmods.refinedstorage.common.support.AbstractBaseContainerMenu;
 import com.refinedmods.refinedstorage.common.support.AbstractBaseScreen;
 import com.refinedmods.refinedstorage.common.support.widget.ScrollbarWidget;
 import com.refinedmods.refinedstorage.common.support.widget.TextMarquee;
+import com.refinedmods.refinedstorage.common.util.ClientPlatformUtil;
 
 import net.minecraft.client.gui.GuiGraphicsExtractor;
 import net.minecraft.client.input.MouseButtonEvent;
@@ -186,7 +187,7 @@ public abstract class AbstractStretchingScreen<T extends AbstractBaseContainerMe
     public boolean mouseScrolled(final double x, final double y, final double z, final double delta) {
         final boolean didScrollbar = scrollbar != null
             && !minecraft.hasShiftDown()
-            && !minecraft.hasControlDown()
+            && !ClientPlatformUtil.isCommandOrControlDown()
             && scrollbar.mouseScrolled(x, y, z, delta);
         return didScrollbar || super.mouseScrolled(x, y, z, delta);
     }