+++ refinedstorage-fabric/src/main/java/com/refinedmods/refinedstorage/fabric/ClientModInitializerImpl.java
@@ -105,6 +105,14 @@ public class ClientModInitializerImpl extends AbstractClientModInitializer imple
                 MenuScreens.register(type, factory::create);
             }
         });
+        // TODO: fix fabric startup warnings
+        // TODO: Fix version info files
+        // TODO: keymappings translation:
+        // "key.category.refinedstorage_quartz_arsenal.keymappings": "Refined Storage - Quartz Arsenal",
+        // TODO: does the tag translation system work?
+        // TODO: Multicraft buttons don't render bg
+        // TODO: Check usages of CTRL (should be CMD)
+        // TODO: scrolling in grid is going wrong direction always down! :)
         registerKeyMappings();
         registerResourceRendering();
         registerAlternativeGridHints();