+++ refinedstorage-neoforge/src/main/java/com/refinedmods/refinedstorage/neoforge/ClientModInitializer.java
@@ -127,6 +127,7 @@ public final class ClientModInitializer extends AbstractClientModInitializer {
 
     @SubscribeEvent
     public static void onRegisterKeyMappings(final RegisterKeyMappingsEvent e) {
+        // TODO: translations.
         final KeyMapping.Category category = new KeyMapping.Category(createIdentifier("keymappings"));
         e.registerCategory(category);