+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/preview/AutocraftingRequestButton.java
@@ -50,6 +50,7 @@ class AutocraftingRequestButton extends AbstractButton {
     @Override
     protected void extractContents(final GuiGraphicsExtractor graphics, final int mouseX, final int mouseY,
                                    final float partialTicks) {
+        extractDefaultSprite(graphics);
         renderResourceIcon(graphics);
         final int yOffset = SmallText.isSmall() ? 8 : 5;
         text.render(graphics, getX() + 3 + 16 + 3, getY() + yOffset, Minecraft.getInstance().font, isHovered);