+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/patterngrid/AlternativesScreen.java
@@ -483,11 +483,11 @@ public class AlternativesScreen extends AbstractAmountScreen<AlternativeContaine
     }
 
     @Override
-    public boolean mouseScrolled(final double x, final double y, final double z, final double delta) {
+    public boolean mouseScrolled(final double x, final double y, final double scrollX, final double scrollY) {
         final boolean didScrollbar = isOverAlternativesArea(x, y)
             && scrollbar != null
-            && scrollbar.mouseScrolled(x, y, z, delta);
-        return didScrollbar || super.mouseScrolled(x, y, z, delta);
+            && scrollbar.mouseScrolled(x, y, scrollX, scrollY);
+        return didScrollbar || super.mouseScrolled(x, y, scrollX, scrollY);
     }
 
     @Override