+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/screen/AbstractGridScreen.java
@@ -555,7 +555,7 @@ public abstract class AbstractGridScreen<T extends AbstractGridContainerMenu> ex
     private boolean canExtract(@Nullable final GridResource resource, final ItemStack carriedStack) {
         return resource != null
             && resource.canExtract(carriedStack, getMenu().getRepository())
-            && !minecraft.hasControlDown();
+            && !ClientPlatformUtil.isCommandOrControlDown();
     }
 
     private boolean canInsert(final int mouseX,
@@ -604,9 +604,8 @@ public abstract class AbstractGridScreen<T extends AbstractGridContainerMenu> ex
     }
 
     @Override
-    public boolean mouseScrolled(final double x, final double y, final double z, final double delta) {
-        final boolean up = delta > 0;
-
+    public boolean mouseScrolled(final double x, final double y, final double scrollX, final double scrollY) {
+        final boolean up = scrollX < 0;
         if (isOverStorageArea((int) x, (int) y)) {
             final GridResource resource = getCurrentGridResource();
             if (resource != null) {
@@ -615,8 +614,7 @@ public abstract class AbstractGridScreen<T extends AbstractGridContainerMenu> ex
         } else if (hoveredSlot != null && hoveredSlot.hasItem() && !(hoveredSlot instanceof DisabledSlot)) {
             mouseScrolledInInventory(up, hoveredSlot);
         }
-
-        return super.mouseScrolled(x, y, z, delta);
+        return super.mouseScrolled(x, y, scrollX, scrollY);
     }
 
     private void mouseScrolledInInventory(final boolean up, final Slot slot) {
@@ -653,11 +651,11 @@ public abstract class AbstractGridScreen<T extends AbstractGridContainerMenu> ex
     @Nullable
     private GridScrollMode getScrollModeWhenScrollingOnGridArea(final boolean up) {
         final boolean shift = minecraft.hasShiftDown();
-        final boolean ctrl = minecraft.hasControlDown();
-        if (shift && ctrl) {
+        final boolean ctrlOrCmd = ClientPlatformUtil.isCommandOrControlDown();
+        if (shift && ctrlOrCmd) {
             return null;
         }
-        return getScrollModeWhenScrollingOnGridArea(up, shift, ctrl);
+        return getScrollModeWhenScrollingOnGridArea(up, shift, ctrlOrCmd);
     }
 
     @Nullable